# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finwiz",
        "description": "Finwiz API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/questions": {
            "get": {
                "tags": [
                    "Question CRUD"
                ],
                "summary": "Get All questions",
                "operationId": "21661f98cc060377f63231be6dec4fc5",
                "responses": {
                    "default": {
                        "description": "Get All questions"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/questions/{question_id}": {
            "get": {
                "tags": [
                    "Question CRUD"
                ],
                "summary": "Get Question",
                "operationId": "140eef689922ccd51cbe293c1b3022ba",
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "path",
                        "description": "Question Id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "integer",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Get Question"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/user/generate-insight": {
            "get": {
                "tags": [
                    "User Transactions Insights"
                ],
                "summary": "Generate user transactions insights",
                "operationId": "05e1031afe05788959204abf053fc727",
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/insights": {
            "get": {
                "tags": [
                    "User Transactions Insights"
                ],
                "summary": "Get user last 6 months transactions insights",
                "operationId": "c290cb4e1471a2e78a6f8a7fc82d053e",
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization ",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization ",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}